Refactor: Break down router/core.py into focused components
Reduces core.py from 1,628 lines to 493 lines by extracting:
- router/algorithms/ - MST, negotiated, and Monte Carlo routing
- router/path.py - Path manipulation and intra-IC routing
- router/adaptive.py - AdaptiveAutorouter and RoutingResult
- router/bus_routing.py - Bus routing integration
- router/diffpair_routing.py - Differential pair routing integration

All existing public APIs maintained for backward compatibility.
All 311 tests pass.

Closes #251
