+++ src/kicad_tools/router/algorithms/__init__.py
@@ -1,0 +1,17 @@
+"""Routing algorithms for the autorouter.
+
+This package contains various routing algorithms:
+- MST (Minimum Spanning Tree) based routing
+- Negotiated congestion routing (PathFinder-style)
+- Monte Carlo multi-start routing
+"""
+
+from .monte_carlo import MonteCarloRouter
+from .mst import MSTRouter
+from .negotiated import NegotiatedRouter
+
+__all__ = [
+    "MSTRouter",
+    "NegotiatedRouter",
+    "MonteCarloRouter",
+]
